build: include own implementation of scalarmult
because we (Thomas and I) are reluctant that it belongs in
tweetnacl. since it breaks the "encryption for dummies" convention in
exposing procedures that "is hard to use". you cannot use scalarmult
as the shared secret directly, for example, because it may reveal
information. ssh does a sha256 over this so we're good in our case.

we'll embed scalarmult* in here for now.