tweak: adjust default channel window size increment and max packet size
because that was partly why it was being slow. chaning the window size
increment to #x800000 takes it from ~500K/s to ~10M/s so that's
better. It's still not 160M/s, I wonder what packet sizes is used in
openssh.

We're using the example same chacha20 C implementation as OpenSSH (I
stole it from there, it's public domain). so the "Scheme overhead"
shouldn't be too big of a deal.

Let's keep this our own little mystery for now.