feat: adds initial support for ssh client in transport layer
simple example provided. this works against my OpenSSH server: I
manage to get the transport layer set up.

next, we need to send a lot of userauth packets and stuff, so a review
of parsing.scm is probably in place to make "unparsing" such packets
easier. maybe some sweet syntax to create a parser and unparser with
one description!