api: add an ssh parameter to unparse- procedures
because I think this will be the most common usecase:

  (write-payload ssh (unparse-channel-data 1 "hello"))

so why not just let users do:

  (unparse-channel-data ssh 1 "hello")

this may be a bad idea. they can still get the raw payload by
substituting #f for ssh. but I guess that'd be mostly debugging
purposes?

unparse-userauth-banner is our first experiment on this.