bugfix: remove incorrect assertion
after running lots of rekeying, we sometimes see this:

    (minissh.scm:448) assertion failed: (>= (string-length K) 32)

this happens because now, with the fixed string->mpint, K (the bignum
sharedsecret) may be smaller than 32 bytes because leading zeros are
removed.

so, in theory, K could be any size between 0 (all zeros!) and
33 (added leading zero) - I think! Anyway, the assertion must go (no,
really, I mean it, it's not just an easy way out this time)!