api major: make read-payload and write-payload thread-safe
these two functions now provide a robust (hopefully) low-level packet
interface for the SSH transport layer, where the encryption and server
authentication is done seamlessly.

making this API thread-safe was hard! but I consider it to be
necessary because it would be really hard to use it single-threaded:
nonblocking read-payload would be very cumbersome I think. also, every
read would potentially write and every write would potentially have to
read because of the internals of the kex phase.

as noted in the disappearing section of the readme, `(kexinit-start)`
is easier to use: it should no longer have any restrictions.

also, note that this also removes the need to `read-payload/nokexinit`
in the kex example. yey! this API makes it possible to call
`(kexinit-start)` safely at any point from any
thread (hopefully!). this will make it easy to respect the 1GiB key
limit.

also, note that this removes the non-blocking read-payload*
version. maybe that was useful? I don't think so.