bugfix: avoid deadlocks when sending after kexinit-start
deadlocks would occur as the sending thread would wait for the read
lock, which was already aquired by the reader thread. the reader
thread would then re-aquire the lock after kexing was done and wait
for the next packet to arrive (which doesn't happen until client
initiates).

the writer packet is still stuck waiting for read lock. this semantics
is incorrect: the writer should only try to lock the read mutex, and
if it fails it means the reader thread is already busy doing the
work. that's fine.

if there isn't a reader thread, though, we need to make sure that
someone is reading and waiting for that kex packet.

I believe the read-packet, write-packet and kexinit-start API is now
thread-safe.