feat: adds userauth-publickey for client public key login
this works but it needs some help. and we really need to settle for a
consistent blob format: keep ssh-ed25519 prefix in our out? maybe
keeping it in will allow us to support more keys in the future. but
then we need to wrap make-asymmetric-sign-keypair to add this
prefix. I don't know if I'm happy about that.

I also don't understand why there's an ssh-ed25519 prefix inside the
key when we already know the key type is ssh-ed25519. maybe there are
key types that don't automatically know the public key format. my gut
is telling me that we should include the ssh-ed25519 _everywhere_.