api major: complete rewrite of minissh-channels
mainly because the API wouldn't really allow for clients. but also
because it was also a bit of a mess. it still is. but now it's a mess
that supports clients.

the new API is a lot less invasive and/or convenient. note the
included tcp-read-timeouts. that's not done for you anymore. and with
the new api you have to loop (with for-each-port for example) and
create threads yourself too. I think this is generally a good thing:
sticking with the single responsibility principle.

the new API adds a `channel-exec` procedure which lets you create
channels, which is the client-side API.

there are new internal workings: a callback system which registers
handlers for valid packet types (eg channel-request is valid when
calling channel-accept, but not necessarily when calling
channel-exec).

each callback/handler processes the incoming packet and shuffles it to
the right gochan, where it can later be processed - or potentially
unblock a thread and be immediately processed. channel-read, for
example will block for incoming channel-data packets (and friends).

it took me really long to settle for a callback system like
this. there is just too much code! I initially wanted it to work with
a single internal buffer, and only processing a set of recognized
packet types there and then, and leave any unhandled ones in the
packet buffer. this made the code _much_ prettier! but it had one
major drawback: it was impossible to know whether an inoming packet
would ever be processed later. some packets (looking at you,
global-request) would never be processed and lie in the packet buffer
forever (causing delays and essentially memory leaks). so no, all
packets have to be processed immediately. some of them will be
buffered away into their own per-channel buffers, others will be
ignored. this means a lot more code :-( but it should be more robust.