cleanup: consistently include algorithm prefix in public keys
any "pk" now means it's an ssh pk which means it contains an alrogithm
prefix (alg-ed25519-add/strip). this is nice for consitency.

kexdh-init/reply now parse as blobs, this means logs no longer contain
binary string which is nice. the parsed payloads also contain the
algorithm prefix (ssh-read-signpk removed it which was confusing).