bugfix: disassociate ssh-channel on channel-close
this is quite tricky. I'm not 100% sure if this is right, but I think
it is. it fixes one problem: remote side closing a channel without us
really sending eof.

if we still have more data to send on that channel, we can do that -
but we'd need to have the channel object since ssh-channel lookup
won't work after we receive 'channel-close.