fix: don't run tests inside with-channel-ports
there sometimes was a deadlock when running the tests. it happend when
the channel thread got the oaat handler and read the next
packet. inside the packet reading, there are tests which would print
to stdout - which again would call channel-write and try to read the
next packet, asking for oaat mutex when we are already inside it.

i can't really put it any clearer than that.