wip major: complete rewrite of channels
because it was horrible. I originally wanted:

1. low level API that let's you send and receive SSH packets without
   thread-safety

2. this turned out to be tricky because if you wanted to respond to
both user input and ssh traffic, you'd need one srfi-18 thread for
reading from the ssh socket and one for the other input sources. so
this low-level API can't really exists.

3. except that a thread-safe low-level API is compeltely horrible (it
still is, this patch doesn't address that (yet)). the problem in
2. should be solved elsewhere. let's settle for a compromise:

in this patch, we try to simplify the channel implementation by
pretending that the ssh packet API is *only* thread-safe if you *only*
receive from one thread and *only* send from another.

this is a step in the right direction, I think. later, minissh.scm can
also be simplified in just supporting partial thread-safety.