log: cleaning up log interface
the logging has now taken a slight turn: it's now geared toward
running an SSH server, where you want to see who tries to log in from
where. the old logging was more of a debugging tool (hence the
current-thread). now the logs look professional and classy with
hardcoded ascii escape sequences for pretty colors which is impossible
to customize for total amazingness.